Made the runtime filter work with the Java snippets, and other snippets...
